+++ src/enums.ts
@@ -9,7 +9,6 @@
 export enum CacheKey {
     BLOCKED_ACCOUNTS = 'BlockedAccounts',
     FAVOURITED_TOOTS = 'FavouritedToots',
-    FAVOURITED_TAG_TOOTS = 'FavouritedHashtagToots',
     FEDIVERSE_POPULAR_SERVERS = 'FediversePopularServers',
     FEDIVERSE_TRENDING_TAGS = 'FediverseTrendingTags',
     FEDIVERSE_TRENDING_LINKS = 'FediverseTrendingLinks',
@@ -20,10 +19,14 @@ export enum CacheKey {
     HOME_TIMELINE_TOOTS = 'HomeTimelineToots',// Toots that the API returns for the home timeline
     MUTED_ACCOUNTS = 'MutedAccounts',
     NOTIFICATIONS = 'Notifications',
-    PARTICIPATED_TAG_TOOTS = 'ParticipatedHashtagToots',
     RECENT_USER_TOOTS = 'RecentUserToots',
     SERVER_SIDE_FILTERS = 'ServerFilters',
     TIMELINE_TOOTS = 'TimelineToots',// The entire timeline (home timeline + trending toots etc.)
+};
+
+export enum TagTootsCacheKey {
+    FAVOURITED_TAG_TOOTS = 'FavouritedHashtagToots',
+    PARTICIPATED_TAG_TOOTS = 'ParticipatedHashtagToots',
     TRENDING_TAG_TOOTS = 'TrendingTagToots'
 };