+++ src/types.ts
@@ -9,12 +9,11 @@ import BooleanFilter, { BooleanFilterArgs, BooleanFilterName } from './filters/b
 import NumericFilter, { NumericFilterArgs } from './filters/numeric_filter';
 import Scorer from './scorer/scorer';
 import Toot, { SerializableToot } from './api/objects/toot';
-import { NonScoreWeightName, ScoreName } from './enums';
-import { CacheKey } from "./enums";
+import { CacheKey, NonScoreWeightName, ScoreName, TagTootsCacheKey } from './enums';
 
 // Records
 export type AccountNames = Record<mastodon.v1.Account["acct"], Account>;
-export type ApiMutex = Record<CacheKey, Mutex>;
+export type ApiMutex = Record<ApiCacheKey, Mutex>;
 export type MastodonInstances = Record<string, MastodonInstance>;
 export type NonScoreWeightInfoDict = Record<NonScoreWeightName, WeightInfo>;
 export type StringDict = Record<string, string>;
@@ -25,6 +24,7 @@ export type WeightInfoDict = Record<WeightName, WeightInfo>;
 
 // Misc
 export type AccountLike = Account | mastodon.v1.Account;
+export type ApiCacheKey = CacheKey | TagTootsCacheKey;
 export type ConcurrencyLockRelease = MutexInterface.Releaser | SemaphoreInterface.Releaser;
 export type CountKey = FilterTitle | string;
 export type FeedFetcher = (api: mastodon.rest.Client) => Promise<Toot[]>;